Add evt build script

+++ scripts/evt_build_centos.sh
@@ -1,13 +1,17 @@
-	OS_VER=$( cat /etc/os-release | grep VERSION_ID | cut -d'=' -f2 | sed 's/[^0-9\.]//gI' | cut -d'.' -f1 )
+	OS_VER=$( cat /etc/os-release | grep VERSION_ID | cut -d'=' -f2 | sed 's/[^0-9\.]//gI' \
+	| cut -d'.' -f1 )
 
 	MEM_MEG=$( free -m | grep Mem | tr -s ' ' | cut -d\  -f2 )
-
 	CPU_SPEED=$( lscpu | grep "MHz" | tr -s ' ' | cut -d\  -f3 | cut -d'.' -f1 )
 	CPU_CORE=$( lscpu | grep "^CPU(s)" | tr -s ' ' | cut -d\  -f2 )
+	MEM_GIG=$(( (($MEM_MEG / 1000) / 2) ))
+	JOBS=$(( ${MEM_GIG} > ${CPU_CORE} ? ${CPU_CORE} : ${MEM_GIG} ))
 
 	DISK_INSTALL=`df -h . | tail -1 | tr -s ' ' | cut -d\  -f1`
-	DISK_TOTAL=`df -h . | tail -1 | tr -s ' ' | cut -d\  -f2 | sed 's/[^0-9\.]//g'`
-	DISK_AVAIL=`df -h . | tail -1 | tr -s ' ' | cut -d\  -f4 | sed 's/[^0-9\.]//g'`
+	DISK_TOTAL_KB=`df . | tail -1 | awk '{print $2}'`
+	DISK_AVAIL_KB=`df . | tail -1 | awk '{print $4}'`
+	DISK_TOTAL=$(( $DISK_TOTAL_KB / 1048576 ))
+	DISK_AVAIL=$(( $DISK_AVAIL_KB / 1048576 ))
 
 	printf "\n\tOS name: $OS_NAME\n"
 	printf "\tOS Version: ${OS_VER}\n"
@@ -18,28 +22,28 @@
 	printf "\tDisk space total: ${DISK_TOTAL%.*}G\n"
 	printf "\tDisk space available: ${DISK_AVAIL%.*}G\n"
 
-	if [ $MEM_MEG -lt 4000 ]; then
-		echo "Your system must have 4 or more Gigabytes of physical memory installed."
+	if [ $MEM_MEG -lt 7000 ]; then
+		echo "Your system must have 7 or more Gigabytes of physical memory installed."
 		echo "exiting now."
 		exit 1
 	fi
 
 	if [ $OS_VER -lt 7 ]; then
-		echo "You must be running Centos 7 or higher to install EOSIO."
-		echo "exiting now"
+		echo "You must be running Centos 7 or higher to install everiToken."
+		echo "exiting now."
 		exit 1
 	fi
 
 	if [ ${DISK_AVAIL%.*} -lt $DISK_MIN ]; then
-		echo "You must have at least ${DISK_MIN}GB of available storage to install EOSIO."
-		echo "exiting now"
+		echo "You must have at least ${DISK_MIN}GB of available storage to install everiToken."
+		echo "exiting now."
 		exit 1
 	fi
 	printf "\n\tChecking Yum installation\n"
 	
 	YUM=$( which yum 2>/dev/null )
 	if [ $? -ne 0 ]; then
-		printf "\n\tYum must be installed to compile EOS.IO.\n"
+		printf "\n\tYum must be installed to compile everiToken.\n"
 		printf "\n\tExiting now.\n"
 		exit 0
 	fi
@@ -49,8 +53,8 @@
 	
 	SCL=$( which scl 2>/dev/null )
 	if [ -z $SCL ]; then
-		printf "\n\tThe Centos Software Collections Repository and devtoolset-7 are required to install EOSIO.\n"
-		printf "\tDo you wish to install and enable this repository and devtoolset package?\n"
+		printf "\n\tThe Centos Software Collections Repository, devtoolset-7 and Python3 are required to install everiToken.\n"
+		printf "\tDo you wish to install and enable this repository, devtoolset-7 and Python3 packages?\n"
 		select yn in "Yes" "No"; do
 			case $yn in
 				[Yy]* ) 
@@ -72,6 +76,15 @@
 					else
 						printf "\n\tCentos devtoolset installed successfully.\n"
 					fi
+					printf "\n\n\tInstalling Python3.\n\n"
+					sudo yum install -y python33.x86_64 2>/dev/null
+					if [ $? -ne 0 ]; then
+						printf "\n\tCentos Python3 installation failed.\n"
+						printf "\n\tExiting now.\n"
+						exit 1
+					else
+						printf "\n\tCentos Python3 installed successfully.\n"
+					fi
 				break;;
 				[Nn]* ) echo "User aborting installation of required Centos Software Collections Repository, Exiting now."; exit;;
 				* ) echo "Please type 1 for yes or 2 for no.";;
@@ -88,7 +101,16 @@
 		printf "\n\tExiting now.\n"
 		exit 1
 	fi
-	printf "\n\tCentos devtoolset-7 successfully enabled.\n"
+	printf "\tCentos devtoolset-7 successfully enabled.\n"
+
+# 	printf "\n\tEnabling Centos python3 installation.\n"
+# 	source /opt/rh/python33/enable
+# 	if [ $? -ne 0 ]; then
+# 		printf "\n\tUnable to enable Centos python3 at this time.\n"
+# 		printf "\n\tExiting now.\n"
+# 		exit 1
+# 	fi
+# 	printf "\tCentos python3 successfully enabled.\n"
 	
 	printf "\n\tUpdating YUM repository.\n"
 
@@ -102,8 +124,8 @@
 
 	printf "\n\tYUM repository successfully updated.\n"
 
-	DEP_ARRAY=( git autoconf automake libtool ocaml.x86_64 doxygen libicu-devel.x86_64 bzip2-devel.x86_64 openssl-devel.x86_64 gmp-devel.x86_64 python-devel.x86_64 gettext-devel.x86_64)
-	DCOUNT=0
+	DEP_ARRAY=( git autoconf automake libtool ocaml.x86_64 doxygen libicu-devel.x86_64 \
+	bzip2-devel.x86_64 openssl-devel.x86_64 gmp-devel.x86_64 python-devel.x86_64 gettext-devel.x86_64)
 	COUNT=1
 	DISPLAY=""
 	DEP=""
@@ -119,15 +141,14 @@
 			DISPLAY="${DISPLAY}${COUNT}. ${DEP_ARRAY[$i]}\n\t"
 			printf "\tPackage ${DEP_ARRAY[$i]} ${bldred} NOT ${txtrst} found.\n"
 			let COUNT++
-			let DCOUNT++
 		else
 			printf "\tPackage ${DEP_ARRAY[$i]} found.\n"
 			continue
 		fi
 	done		
 
-	if [ ${DCOUNT} -ne 0 ]; then
-		printf "\n\tThe following dependencies are required to install EOSIO.\n"
+	if [ ${COUNT} -gt 1 ]; then
+		printf "\n\tThe following dependencies are required to install everiToken.\n"
 		printf "\n\t$DISPLAY\n\n"
 		printf "\tDo you wish to install these dependencies?\n"
 		select yn in "Yes" "No"; do
@@ -151,16 +172,20 @@
 		printf "\n\tNo required YUM dependencies to install.\n"
 	fi
 
-	printf "\n\tChecking for CMAKE.\n"
-    # install CMAKE 3.10.2
+	printf "\n\tChecking CMAKE installation.\n"
     if [ ! -e ${CMAKE} ]; then
 		printf "\tInstalling CMAKE\n"
 		mkdir -p ${HOME}/opt/ 2>/dev/null
 		cd ${HOME}/opt
-		curl -L -O https://cmake.org/files/v3.10/cmake-3.10.2.tar.gz
+		STATUS=$(curl -LO -w '%{http_code}' --connect-timeout 30 https://cmake.org/files/v3.10/cmake-3.10.2.tar.gz)
+		if [ "${STATUS}" -ne 200 ]; then
+			printf "\tUnable to download CMAKE at this time.\n"
+			printf "\tExiting now.\n\n"
+			exit;
+		fi
 		tar xf cmake-3.10.2.tar.gz
 		rm -f cmake-3.10.2.tar.gz
-		ln -s cmake-3.10.2/ cmake
+		ln -s ${HOME}/opt/cmake-3.10.2/ cmake
 		cd cmake
 		./bootstrap
 		if [ $? -ne 0 ]; then
@@ -168,7 +193,7 @@
 			printf "\tExiting now.\n\n"
 			exit;
 		fi
-		make
+		make -j${JOBS}
 		if [ $? -ne 0 ]; then
 			printf "\tError compiling CMAKE.\n"
 			printf "\tExiting now.\n\n"
@@ -178,28 +203,55 @@
 		printf "\tCMAKE found\n"
 	fi
 
-	printf "\n\tChecking for boost libraries\n"
-	if [ ! -d ${HOME}/opt/boost_1_66_0 ]; then
-		# install boost
-		printf "\tInstalling boost libraries\n"
+	printf "\n\tChecking boost library installation.\n"
+	BVERSION=`cat "${BOOST_ROOT}/include/boost/version.hpp" 2>/dev/null | grep BOOST_LIB_VERSION \
+	| tail -1 | tr -s ' ' | cut -d\  -f3 | sed 's/[^0-9\._]//gI'`
+	if [ "${BVERSION}" != "1_66" ]; then
+		printf "\tRemoving existing boost libraries in ${HOME}/opt/boost* .\n"
+		rm -rf ${HOME}/opt/boost*
+		if [ $? -ne 0 ]; then
+			printf "\n\tUnable to remove deprecated boost libraries at this time.\n"
+			printf "\n\tExiting now.\n"
+			exit 1
+		fi
+		printf "\tInstalling boost libraries.\n"
 		cd ${TEMP_DIR}
-		curl -L https://dl.bintray.com/boostorg/release/1.66.0/source/boost_1_66_0.tar.bz2 > boost_1.66.0.tar.bz2
-		tar xf boost_1.66.0.tar.bz2
-		cd boost_1_66_0/
+		STATUS=$(curl -LO -w '%{http_code}' --connect-timeout 30 https://dl.bintray.com/boostorg/release/1.66.0/source/boost_1_66_0.tar.bz2)
+		if [ "${STATUS}" -ne 200 ]; then
+			printf "\tUnable to download Boost libraries at this time.\n"
+			printf "\tExiting now.\n\n"
+			exit;
+		fi
+		tar xf ${TEMP_DIR}/boost_1_66_0.tar.bz2
+		rm -f  ${TEMP_DIR}/boost_1_66_0.tar.bz2
+		cd ${TEMP_DIR}/boost_1_66_0/
 		./bootstrap.sh "--prefix=$BOOST_ROOT"
+		if [ $? -ne 0 ]; then
+			printf "\n\tInstallation of boost libraries failed. 0\n"
+			printf "\n\tExiting now.\n"
+			exit 1
+		fi
 		./b2 install
+		if [ $? -ne 0 ]; then
+			printf "\n\tInstallation of boost libraries failed. 1\n"
+			printf "\n\tExiting now.\n"
+			exit 1
+		fi
 		rm -rf ${TEMP_DIR}/boost_1_66_0/
-		rm -f  ${TEMP_DIR}/boost_1.66.0.tar.bz2
 	else
-		printf "\tBoost 1.66 found at ${HOME}/opt/boost_1_66_0\n"
+		printf "\tBoost 1.66.0 found at ${HOME}/opt/boost_1_66_0.\n"
 	fi
 
-	printf "\n\tChecking for secp256k1-zkp\n"
-    # install secp256k1-zkp (Cryptonomex branch)
+	printf "\n\tChecking secp256k1-zkp installation.\n"
     if [ ! -e /usr/local/lib/libsecp256k1.a ]; then
 		printf "\tInstalling secp256k1-zkp (Cryptonomex branch)\n"
 		cd ${TEMP_DIR}
 		git clone https://github.com/cryptonomex/secp256k1-zkp.git
+		if [ $? -ne 0 ]; then
+			printf "\tUnable to clone repo secp256k1-zkp @ https://github.com/cryptonomex/secp256k1-zkp.git.\n"
+			printf "\tExiting now.\n\n"
+			exit;
+		fi
 		cd secp256k1-zkp
 		./autogen.sh
 		if [ $? -ne 0 ]; then
@@ -208,7 +260,7 @@
 			exit;
 		fi
 		./configure
-		make
+		make -j${JOBS}
 		if [ $? -ne 0 ]; then
 			printf "\tError compiling secp256k1-zkp.\n"
 			printf "\tExiting now.\n\n"
@@ -220,56 +272,9 @@
 		printf "\tsecp256k1 found\n"
 	fi
 	
-	printf "\n\tChecking for binaryen\n"
-	if [ ! -d ${HOME}/opt/binaryen ]; then
-		# Install binaryen v1.37.14:
-		printf "\tInstalling binaryen v1.37.14:\n"
-		cd ${TEMP_DIR}
-		git clone https://github.com/EOSIO/binaryen
-		cd binaryen
-		git checkout eosio
-		$CMAKE . && make
-		if [ $? -ne 0 ]; then
-			printf "\tError compiling binaryen.\n"
-			printf "\tExiting now.\n\n"
-			exit;
-		fi
-		mkdir -p ${HOME}/opt/binaryen/ 2>/dev/null
-		mv ${TEMP_DIR}/binaryen/bin ${HOME}/opt/binaryen/
-		rm -rf ${TEMP_DIR}/binaryen
-	else
-		printf "\tBinaryen found at ${HOME}/opt/binaryen\n"
-	fi
-
-	printf "\n\tChecking for LLVM with WASM support.\n"
-	if [ ! -d ${HOME}/opt/wasm/bin ]; then
-		# Build LLVM and clang with EXPERIMENTAL WASM support:
-		printf "\tInstalling LLVM & WASM\n"
-		cd ${TEMP_DIR}
-		mkdir llvm-compiler  2>/dev/null
-		cd llvm-compiler
-		git clone --depth 1 --single-branch --branch release_40 https://github.com/llvm-mirror/llvm.git
-		cd llvm/tools
-		git clone --depth 1 --single-branch --branch release_40 https://github.com/llvm-mirror/clang.git
-		cd ..
-		mkdir build 2>/dev/null
-		cd build
-		$CMAKE -G "Unix Makefiles" -DCMAKE_INSTALL_PREFIX=${HOME}/opt/wasm \
-		-DLLVM_TARGETS_TO_BUILD="host" -DLLVM_EXPERIMENTAL_TARGETS_TO_BUILD=WebAssembly \
-		-DLLVM_ENABLE_RTTI=1 -DCMAKE_BUILD_TYPE=Release ../
-		if [ $? -ne 0 ]; then
-			printf "\tError compiling LLVM and clang with EXPERIMENTAL WASM support.\n"
-			printf "\tExiting now.\n\n"
-			exit;
-		fi
-		make -j$(nproc)
-		if [ $? -ne 0 ]; then
-			printf "\tError compiling LLVM and clang with EXPERIMENTAL WASM support.\n"
-			printf "\tExiting now.\n\n"
-			exit;
-		fi
-		make install
-		rm -rf ${TEMP_DIR}/llvm-compiler 2>/dev/null
-	else
-		printf "\tWASM found at ${HOME}/opt/wasm\n"
-	fi
+	function print_instructions()
+	{
+		printf "\tsource /opt/rh/python33/enable\n"
+		printf "\tcd ${HOME}/evt/build; make test\n\n"
+	return 0
+	}
